MacroDecompose.jl is linked to GenX-Benders
MacroEnergySystemsDecomposition.jl is linked in GenX-Benders.jl. The algorithms in MacroDecomp.jl can be used in GenX-Benders.ji to solve GenX instances. The work flow is: 1. make changes in MacroDecomp.jl and do git push, 2. update the packages in GenX-Benders.jl, 3. run the instances in GenX-Benders.jl. However, currently, the cases cannot be directly run in MacroDecomp.jl.